requirements.txt
discord.py
python-dotenv
yfinance
pandas